frontend/requirements.txt
# Core Dependencies
streamlit>=1.32.0,<2.0.0
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0

# API Client
requests>=2.31.0,<3.0.0
httpx[http2]>=0.24.0,<1.0.0

# Data Processing
python-dotenv>=1.0.0,<2.0.0
python-multipart>=0.0.6,<1.0.0
Pillow>=10.0.0,<11.0.0
PyPDF2>=3.0.0,<4.0.0
python-docx>=1.1.0,<2.0.0

# Authentication
python-jose[cryptography]>=3.3.0,<4.0.0
python-keycloak>=3.6.0,<4.0.0
python-jwt>=4.0.0,<5.0.0
supabase>=2.0.0,<3.0.0

# Utilities
python-dateutil>=2.8.2,<3.0.0
pytz>=2023.3,<2024.0
tzlocal>=5.0.1,<6.0.0
pyperclip>=1.8.2,<2.0.0

# Monitoring and Logging
sentry-sdk>=1.30.0,<2.0.0
python-json-logger>=2.0.7,<3.0.0

# Testing (development only)
pytest>=7.4.0,<8.0.0
pytest-cov>=4.1.0,<5.0.0
pytest-mock>=3.11.1,<4.0.0
pytest-asyncio>=0.21.1,<1.0.0

# Code Quality (development only)
black>=23.7.0,<24.0.0
isort>=5.12.0,<6.0.0
flake8>=6.1.0,<7.0.0
mypy>=1.5.0,<2.0.0

# Documentation (development only)
mkdocs>=1.5.0,<2.0.0
mkdocs-material>=9.1.0,<10.0.0
mkdocstrings[python]>=0.22.0,<1.0.0
